Move all result data trasnformation into Trace class
